+++ hpOneView/connection.py
@@ -439,7 +439,12 @@ class connection(object):
 
         self._cred = cred
         try:
-            task, body = self.post(uri['loginSessions'], self._cred)
+            if self._cred.get("sessionID"):
+                self.set_session_id(self._cred["sessionID"])
+                task, body = self.put(uri['loginSessions'], None)
+            else:
+                self._cred.pop("sessionID", None)
+                task, body = self.post(uri['loginSessions'], self._cred)
         except HPOneViewException:
             logger.exception('Login failed')
             raise