+++ tests/unit/test_oneview_client.py
@@ -86,6 +86,11 @@ OS_ENVIRON_CONFIG_MINIMAL = {
     'ONEVIEWSDK_PASSWORD': 'secret123'
 }
 
+OS_ENVIRON_CONFIG_MINIMAL_WITH_SESSIONID = {
+    'ONEVIEWSDK_IP': '172.16.100.199',
+    'ONEVIEWSDK_SESSIONID': '123'
+}
+
 OS_ENVIRON_CONFIG_FULL = {
     'ONEVIEWSDK_IP': '172.16.100.199',
     'ONEVIEWSDK_IMAGE_STREAMER_IP': '172.172.172.172',
@@ -93,6 +98,16 @@ OS_ENVIRON_CONFIG_FULL = {
     'ONEVIEWSDK_PASSWORD': 'secret123',
     'ONEVIEWSDK_API_VERSION': '201',
     'ONEVIEWSDK_AUTH_LOGIN_DOMAIN': 'authdomain',
+    'ONEVIEWSDK_PROXY': '172.16.100.195:9999'
+}
+
+OS_ENVIRON_CONFIG_FULL_WITH_SESSIONID = {
+    'ONEVIEWSDK_IP': '172.16.100.199',
+    'ONEVIEWSDK_IMAGE_STREAMER_IP': '172.172.172.172',
+    'ONEVIEWSDK_USERNAME': 'admin',
+    'ONEVIEWSDK_PASSWORD': 'secret123',
+    'ONEVIEWSDK_SESSIONID': '123',
+    'ONEVIEWSDK_API_VERSION': '201',
     'ONEVIEWSDK_PROXY': '172.16.100.195:9999'
 }
 
@@ -149,6 +164,39 @@ class OneViewClientTest(unittest.TestCase):
 
     @mock.patch.object(connection, 'login')
     @mock.patch(mock_builtin('open'))
+    def test_from_json_file_with_sessionID(self, mock_open, mock_login):
+        json_config_content = u"""{
+          "ip": "172.16.102.59",
+          "credentials": {
+            "userName": "administrator",
+            "authLoginDomain": "",
+            "password": "",
+            "sessionID": "123"
+          }
+        }"""
+        mock_open.return_value = self.__mock_file_open(json_config_content)
+        oneview_client = OneViewClient.from_json_file("config.json")
+
+        self.assertIsInstance(oneview_client, OneViewClient)
+        self.assertEqual("172.16.102.59", oneview_client.connection.get_host())
+
+    @mock.patch.object(connection, 'login')
+    @mock.patch(mock_builtin('open'))
+    def test_from_json_file_with_only_sessionID(self, mock_open, mock_login):
+        json_config_content = u"""{
+          "ip": "172.16.102.59",
+          "credentials": {
+            "sessionID": "123"
+          }
+        }"""
+        mock_open.return_value = self.__mock_file_open(json_config_content)
+        oneview_client = OneViewClient.from_json_file("config.json")
+
+        self.assertIsInstance(oneview_client, OneViewClient)
+        self.assertEqual("172.16.102.59", oneview_client.connection.get_host())
+
+    @mock.patch.object(connection, 'login')
+    @mock.patch(mock_builtin('open'))
     def test_default_api_version(self, mock_open, mock_login):
         json_config_content = u"""{
           "ip": "172.16.102.59",
@@ -190,7 +238,21 @@ class OneViewClientTest(unittest.TestCase):
 
         mock_login.assert_called_once_with(dict(userName='admin',
                                                 password='secret123',
-                                                authLoginDomain=''))
+                                                authLoginDomain='',
+                                                sessionID=''))
+        mock_set_proxy.assert_not_called()
+        self.assertEqual(300, oneview_client.connection._apiVersion)
+
+    @mock.patch.object(connection, 'login')
+    @mock.patch.object(connection, 'set_proxy')
+    @mock.patch.dict('os.environ', OS_ENVIRON_CONFIG_MINIMAL_WITH_SESSIONID)
+    def test_from_minimal_environment_variables_with_sessionID(self, mock_set_proxy, mock_login):
+        oneview_client = OneViewClient.from_environment_variables()
+
+        mock_login.assert_called_once_with(dict(userName='',
+                                                password='',
+                                                authLoginDomain='',
+                                                sessionID='123'))
         mock_set_proxy.assert_not_called()
         self.assertEqual(300, oneview_client.connection._apiVersion)
 
@@ -202,12 +264,29 @@ class OneViewClientTest(unittest.TestCase):
 
         mock_login.assert_called_once_with(dict(userName='admin',
                                                 password='secret123',
-                                                authLoginDomain='authdomain'))
+                                                authLoginDomain='authdomain',
+                                                sessionID=''))
         mock_set_proxy.assert_called_once_with('172.16.100.195', 9999)
 
         self.assertEqual(201, oneview_client.connection._apiVersion)
         self.assertEqual(oneview_client.create_image_streamer_client().connection.get_host(),
                          OS_ENVIRON_CONFIG_FULL['ONEVIEWSDK_IMAGE_STREAMER_IP'])
+
+    @mock.patch.object(connection, 'login')
+    @mock.patch.object(connection, 'set_proxy')
+    @mock.patch.dict('os.environ', OS_ENVIRON_CONFIG_FULL_WITH_SESSIONID)
+    def test_from_full_environment_variables_with_sessionID(self, mock_set_proxy, mock_login):
+        oneview_client = OneViewClient.from_environment_variables()
+
+        mock_login.assert_called_once_with(dict(userName='admin',
+                                                password='secret123',
+                                                authLoginDomain='',
+                                                sessionID='123'))
+        mock_set_proxy.assert_called_once_with('172.16.100.195', 9999)
+
+        self.assertEqual(201, oneview_client.connection._apiVersion)
+        self.assertEqual(oneview_client.create_image_streamer_client().connection.get_host(),
+                         OS_ENVIRON_CONFIG_FULL_WITH_SESSIONID['ONEVIEWSDK_IMAGE_STREAMER_IP'])
 
     @mock.patch.dict('os.environ', OS_ENVIRON_CONFIG_FULL)
     @mock.patch.object(OneViewClient, '__init__')
@@ -219,9 +298,40 @@ class OneViewClientTest(unittest.TestCase):
                                           'ip': '172.16.100.199',
                                           'image_streamer_ip': '172.172.172.172',
                                           'credentials':
-                                              {'password': 'secret123',
+                                              {'userName': 'admin',
+                                               'password': 'secret123',
                                                'authLoginDomain': 'authdomain',
-                                               'userName': 'admin'}})
+                                               'sessionID': ''}})
+
+    @mock.patch.dict('os.environ', OS_ENVIRON_CONFIG_FULL_WITH_SESSIONID)
+    @mock.patch.object(OneViewClient, '__init__')
+    def test_from_environment_variables_is_passing_right_arguments_to_the_constructor_with_sessionID(self, mock_cls):
+        mock_cls.return_value = None
+        OneViewClient.from_environment_variables()
+        mock_cls.assert_called_once_with({'api_version': 201,
+                                          'proxy': '172.16.100.195:9999',
+                                          'ip': '172.16.100.199',
+                                          'image_streamer_ip': '172.172.172.172',
+                                          'credentials':
+                                              {'userName': 'admin',
+                                               'password': 'secret123',
+                                               'authLoginDomain': '',
+                                               'sessionID': '123'}})
+
+    @mock.patch.dict('os.environ', OS_ENVIRON_CONFIG_MINIMAL_WITH_SESSIONID)
+    @mock.patch.object(OneViewClient, '__init__')
+    def test_from_environment_variables_is_passing_right_arguments_to_the_constructor_with_only_sessionID(self, mock_cls):
+        mock_cls.return_value = None
+        OneViewClient.from_environment_variables()
+        mock_cls.assert_called_once_with({'api_version': 300,
+                                          'proxy': '',
+                                          'ip': '172.16.100.199',
+                                          'image_streamer_ip': '',
+                                          'credentials':
+                                              {'userName': '',
+                                               'password': '',
+                                               'authLoginDomain': '',
+                                               'sessionID': '123'}})
 
     @mock.patch.object(connection, 'login')
     def test_create_image_streamer_client_without_image_streamer_ip(self, mock_login):